requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
cloudscraper==1.2.60
DateTime==4.5
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
py-cord==2.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
six==1.16.0
urllib3==1.26.11
yarl==1.8.1
zope.interface==5.4.0